refactor: Merge util package into endpoint package
