+++ config/endpoint/key.go
@@ -1,4 +1,4 @@
-package util
+package endpoint
 
 import "strings"
