+++ config/endpoint/key_bench_test.go
@@ -1,4 +1,4 @@
-package util
+package endpoint
 
 import (
 	"testing"
