+++ config/endpoint/key_test.go
@@ -1,4 +1,4 @@
-package util
+package endpoint
 
 import "testing"
