+++ config/endpoint/status.go
@@ -1,9 +1,5 @@
 package endpoint
 
-import (
-	"github.com/TwiN/gatus/v5/util"
-)
-
 // Status contains the evaluation Results of an Endpoint
 type Status struct {
 	// Name of the endpoint
@@ -34,7 +30,7 @@ func NewStatus(group, name string) *Status {
 	return &Status{
 		Name:    name,
 		Group:   group,
-		Key:     util.ConvertGroupAndEndpointNameToKey(group, name),
+		Key:     ConvertGroupAndEndpointNameToKey(group, name),
 		Results: make([]*Result, 0),
 		Events:  make([]*Event, 0),
 		Uptime:  NewUptime(),
